+++ zanata-war/src/main/java/org/zanata/rest/service/TranslationMemoryResourceService.java
@@ -100,6 +100,7 @@ public class TranslationMemoryResourceService implements
     }
 
     @Override
+    @Restrict("#{s:hasPermission('', 'download-tmx')}")
     public Response getProjectTranslationMemory(@Nonnull String projectSlug,
             @Nullable LocaleId locale) {
         log.debug("exporting TMX for project {}, locale {}", projectSlug,
@@ -118,6 +119,7 @@ public class TranslationMemoryResourceService implements
     }
 
     @Override
+    @Restrict("#{s:hasPermission('', 'download-tmx')}")
     public Response getProjectIterationTranslationMemory(
             @Nonnull String projectSlug, @Nonnull String iterationSlug,
             @Nullable LocaleId locale) {